Add WormTraitPanel component - rename ConfigurableParam to ConfigurableTrait
